+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 0.8.0 (Unreleased)
   * Merge declarator constructors. Now you differentiate between array and
     scalar declarators by looking at the relevant field.
+  * Add `bozAsNatural :: Num a => Boz -> a` function to resolve BOZ constants as
+    integers
 
 ### 0.7.0 (Dec 09, 2021)
   * No longer treat `!` in strings as comments in continuation reformatter